refactor: Domain 관련 String에서 enum으로 수정
