+++ lovebird-api/src/test/kotlin/com/lovebird/api/common/base/ControllerDescribeSpec.kt
@@ -1,41 +1,38 @@
 package com.lovebird.api.common.describe
 
 import com.fasterxml.jackson.databind.ObjectMapper
+import com.lovebird.api.config.WebMvcConfig
 import com.lovebird.api.controller.external.PresignedUrlController
 import com.lovebird.api.utils.restdocs.restDocMockMvcBuild
+import io.kotest.core.spec.BeforeEach
+import io.kotest.core.spec.BeforeTest
 import io.kotest.core.spec.style.DescribeSpec
+import org.junit.jupiter.api.extension.ExtendWith
 import org.mockito.Mockito
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest
+import org.springframework.context.annotation.Import
 import org.springframework.restdocs.ManualRestDocumentation
+import org.springframework.restdocs.RestDocumentationExtension
 import org.springframework.test.web.servlet.MockMvc
 import org.springframework.web.context.WebApplicationContext
 
+@Import(WebMvcConfig::class)
+@ExtendWith(RestDocumentationExtension::class)
 @WebMvcTest(
 	PresignedUrlController::class
 )
 abstract class ControllerDescribeSpec(
-	@Autowired
-	private val context: WebApplicationContext,
-	body: DescribeSpec.() -> Unit = {}
+	body: DescribeSpec.() -> Unit = {},
 ) : DescribeSpec(body) {
 
-	@Autowired
-	protected val mapper: ObjectMapper
+	companion object {
+		private val mapper: ObjectMapper = ObjectMapper()
 
-	@Autowired
-	protected val mockMvc: MockMvc
+		fun toJson(value: Any): String {
+			return mapper.writeValueAsString(value)
+		}
 
-	protected val restDocumentation: ManualRestDocumentation = ManualRestDocumentation()
-
-	init {
-		mockMvc = restDocMockMvcBuild(context, restDocumentation)
-		mapper = ObjectMapper()
+		fun <T> any(type: Class<T>): T = Mockito.any<T>(type)
 	}
-
-	protected fun toJson(value: Any): String {
-		return mapper.writeValueAsString(value)
-	}
-
-	protected fun <T> any(type: Class<T>): T = Mockito.any<T>(type)
 }
