+++ lovebird-common/src/main/kotlin/com/lovebird/common/enums/Domain.kt
@@ -1,0 +1,11 @@
+package com.lovebird.common.enums
+
+enum class Domain {
+
+	PROFILE,
+	DIARY;
+
+	fun lower(): String {
+		return this.name.lowercase()
+	}
+}
